Added Company ID to HubSpotApi and created Demo API for entities
In this commit, a Company ID was added to the existing HubSpotObjectIds class in the HubSpotApi. Furthermore, a new DemonstrateApi class was created under HubSpot.NET.Examples which contains example code for creating and deleting entities, such as a new Contact and Company. This provides clear and practical examples for users on how to use the API functionalities.