Replace CreateTestContact method with RecreateTestContact
The CreateTestContact method in the HubSpotIntegrationTestBase.cs file was removed and all its calls were replaced with RecreateTestContact in both HubSpotCompanyApiIntegrationTests.cs and HubSpotContactApiIntegrationTests.cs files. The aim of this change is to only use a single method for creating test contacts in order to make the tests more reliable and consistent.