Update project details in csproj file
The existing project details in the csproj file are updated. This includes the addition of "PowerDiary" to the authors and copyright details and changes to the company name and PackageId to reflect the new contributing organization, "PowerDiary". The description of the project is also changed to more accurately represent its purpose.